Little changes for types

+++ sort/insertion_sort/is_1.c
@@ -31,7 +31,7 @@ void i_sort(size_t arr[], size_t size)
 		key = arr[i];
 		j = i - 1;
 
-		while(j >= 0 && arr[j] > key)
+		while(j >= 0 && arr[j] > (size_t)key)
 		{
 			arr[j + 1] = arr[j];
 			--j;
